Reduce scope of edit response
