+++ src/queries/responses_ollama.hpp
@@ -16,21 +16,4 @@ struct OllamaResponse: public SuccessResponse {
     OllamaResponse(const std::string &response);
 };
 
-class OllamaEdit: public OllamaResponse {
-public:
-    OllamaEdit() :
-        OllamaResponse() {} // needed for variants
-    OllamaEdit(const std::string &response, const double total_t);
-
-    std::string description_of_changes;
-    std::string code;
-
-    double total_time = 0.0;
-    int input_tokens = 0;
-    int output_tokens = 0;
-
-private:
-    void unpack_structured_output_();
-};
-
 } // namespace queries
